moved enum things around
