+++ backend/src/models/bid.py
@@ -1,60 +1,9 @@
 from datetime import datetime
 from . import db
-from enum import Enum
 
-# IDK if this is the standart
+# IDK if this is the standard
 from models.round import Round
 
-
-class Bets(Enum):
-    ONE = 1
-    TWO = 2
-    THREE = 3
-    FOUR = 4
-    FIVE = 5
-    SIX = 6
-    SEVEN = 7
-    EIGHT = 8
-    NINE = 9
-    TEN = 10
-    ELEVEN = 11
-    TWELVE = 12
-    THIRTEEN = 13
-    FOURTEEN = 14
-    FIFTEEN = 15
-    SIXTEEN = 16
-    SEVENTEEN = 17
-    EIGHTEEN = 18
-    NINETEEN = 19
-    TWENTY = 20
-    TWENTY_ONE = 21
-    TWENTY_TWO = 22
-    TWENTY_THREE = 23
-    TWENTY_FOUR = 24
-    TWENTY_FIVE = 25
-    TWENTY_SIX = 26
-    TWENTY_SEVEN = 27
-    TWENTY_EIGHT = 28
-    TWENTY_NINE = 29
-    THIRTY = 30
-    THIRTY_ONE = 31
-    THIRTY_TWO = 32
-    THIRTY_THREE = 33
-    THIRTY_FOUR = 34
-    THIRTY_FIVE = 35
-    THIRTY_SIX = 36
-    HALF_ONE = 37  # 1-18
-    HALF_TWO = 38  # 19-36
-    EVEN = 39  # mod 2 = 0
-    ODD = 40  # mod 2 = 1
-    BLACK = 41  # 2-4-6-8-10-11-13-15-17-20-22-24-26-28-29-31-33-35
-    RED = 42  #  1-3-5-7-9-12-14-16-18-19-21-23-25-27-30-32-34-36
-    THIRD_ONE = 43  # 1-12
-    THIRD_TWO = 44  # 13-24
-    THIRD_THREE = 45  # 25-36
-    ROW_ONE = 46  # mod 3 = 1
-    ROW_TWO = 47  # mod 3 = 2
-    ROW_THREE = 48  # mod 3 = 0
 
 
 
@@ -90,7 +39,7 @@ class Bid(db.Model):
 
     @classmethod
     def new(cls, wager, bet, user_id, round_id):
-        n = cls(wager=wager, bet=bet, user_id=user_id, round_id=round_id, game_id=Round.get())
+        n = cls(wager=wager, bet=bet, user_id=user_id, round_id=round_id, game_id=Round.get()) # TODO
         db.session.add(n)
         db.session.commit()
         return n
